+++ tests/test_ut_database.py
@@ -17,30 +17,30 @@
 #  You should have received a copy of the GNU Affero General Public License
 #  along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
-"""test_ut_sqlitedb: Test Sqlitedb() class."""
+"""test_ut_database: Test Database() class."""
 
 from snippy.constants import Constants as Const
 from tests.testlib.content import Content
-from tests.testlib.database import Database
+from tests.testlib.database import Database as TDatabase
 from tests.testlib.snippet import Snippet
 
 
-class TestUtSqlitedb(object):
-    """Test Sqlitedb() class."""
+class TestUtDatabase(object):
+    """Test Database() class."""
 
-    def test_sqlitedb_insert_001(self, sqlitedb, cause):
-        """Test SqliteDb basic insert.
+    def test_database_insert_001(self, database, cause):
+        """Test database basic insert.
 
         Insert one Snippet resource into the database.
         """
 
         collection = Content.get_collection(Snippet.REMOVE)
-        sqlitedb.insert(collection)
+        database.insert(collection)
         cause.assert_called_once_with('201 Created', 'content created')
-        assert collection == Database.get_snippets()
-        assert len(Database.get_snippets()) == 1
+        assert collection == TDatabase.get_snippets()
+        assert len(TDatabase.get_snippets()) == 1
 
-    def test_sqlitedb_insert_002(self, sqlitedb, cause):
+    def test_database_insert_002(self, database, cause):
         """Test SqliteDb basic insert.
 
         Insert four Snippet resources into the database. This verifies with
@@ -51,56 +51,56 @@ class TestUtSqlitedb(object):
         collection.migrate(Content.get_collection(Snippet.FORCED))
         collection.migrate(Content.get_collection(Snippet.EXITED))
         collection.migrate(Content.get_collection(Snippet.NETCAT))
-        sqlitedb.insert(collection)
+        database.insert(collection)
         cause.assert_called_once_with('201 Created', 'content created')
-        assert collection == Database.get_snippets()
-        assert len(Database.get_snippets()) == 4
+        assert collection == TDatabase.get_snippets()
+        assert len(TDatabase.get_snippets()) == 4
 
-    def test_sqlitedb_select_001(self, sqlitedb, cause):
-        """Test SqliteDb basic select.
+    def test_database_select_001(self, database, cause):
+        """Test database basic select.
 
-        Select content with regexp stored into sqlite. In this case the last
+        Select content with regexp stored into database. In this case the last
         keyword matches to links column.
         """
 
         collection = Content.get_collection(Snippet.FORCED)
         collection.migrate(Content.get_collection(Snippet.EXITED))
         collection.migrate(Content.get_collection(Snippet.NETCAT))
-        sqlitedb.insert(collection)
-        collection = sqlitedb.select(scat=(Const.SNIPPET,), sall=('foo', 'bar', 'digitalocean'))
+        database.insert(collection)
+        collection = database.select(scat=(Const.SNIPPET,), sall=('foo', 'bar', 'digitalocean'))
         cause.assert_called_once_with('201 Created', 'content created')
         assert collection == Content.get_collection(Snippet.FORCED)
 
-    def test_sqlitedb_delete_002(self, sqlitedb, cause, mocker):
-        """Test SqliteDb basic delete.
+    def test_database_delete_002(self, database, cause, mocker):
+        """Test database basic delete.
 
         Delete one row from database with short digest.
         """
 
         collection = Content.get_collection(Snippet.REMOVE)
         collection.migrate(Content.get_collection(Snippet.FORCED))
-        sqlitedb.insert(collection)
-        sqlitedb.delete('53908d68425c61dc')
+        database.insert(collection)
+        database.delete('53908d68425c61dc')
         results = []
         results.append(mocker.call('201 Created', 'content created'))
         results.append(mocker.call('204 No Content', 'content deleted successfully'))
         cause.assert_has_calls(results)
-        assert Database.get_snippets() == Content.get_collection(Snippet.REMOVE)
-        assert len(Database.get_snippets()) == 1
+        assert TDatabase.get_snippets() == Content.get_collection(Snippet.REMOVE)
+        assert len(TDatabase.get_snippets()) == 1
 
-    def test_sqlitedb_delete_001(self, sqlitedb, cause, mocker):
-        """Test SqliteDb basic delete.
+    def test_database_delete_001(self, database, cause, mocker):
+        """Test database basic delete.
 
         Delete one row from database with long digest.
         """
 
         collection = Content.get_collection(Snippet.REMOVE)
         collection.migrate(Content.get_collection(Snippet.FORCED))
-        sqlitedb.insert(collection)
-        sqlitedb.delete('53908d68425c61dc310c9ce49d530bd858c5be197990491ca20dbe888e6deac5')
+        database.insert(collection)
+        database.delete('53908d68425c61dc310c9ce49d530bd858c5be197990491ca20dbe888e6deac5')
         results = []
         results.append(mocker.call('201 Created', 'content created'))
         results.append(mocker.call('204 No Content', 'content deleted successfully'))
         cause.assert_has_calls(results)
-        assert Database.get_snippets() == Content.get_collection(Snippet.REMOVE)
-        assert len(Database.get_snippets()) == 1
+        assert TDatabase.get_snippets() == Content.get_collection(Snippet.REMOVE)
+        assert len(TDatabase.get_snippets()) == 1